feat(schema) accept custom 'err' for 'eq' and 'ne' validators
- Add support for custom error message in fields schema
- Add support for returning said custom message upon `eq`/`ne` validation
failure

While the new `err` field is only leveraged in the `eq` and `ne` validators,
other validator checks can be extended to follow similar pattern. Doing so
allows schemas to provide friendlier error messages without the need for
additional entity checkers.

A subsequent commit will build on top of this commit to simplify routes
subschemas, by replacing a number of conditional checkers (with
`then_err` fields) with an `err` field in the field's schema:

```
    { conditional = { if_field = "protocols",
                      if_match = { elements = { type = "string", one_of = { "http", "https" }}},
                      then_field = "destinations",
                      then_match = { len_eq = 0 },
                      then_err = "cannot set 'destinations' when 'protocols' is 'http' or 'https'",
                    }},
```

Given this conditional check's purpose is to ensure the `destinations`
field isn't set for the entity, it can be replaced with the following,
leveraging the new `err` field:

```
    { destinations = typedefs.no_destinations {
        err = "cannot set 'destinations' when 'protocols' is 'http' or 'https'"
    }}
```